clean up newly added matrix decomposition implementations
