+++ examples/shackleton_vary_outer/collect_time_dep_data_long_spinup_phase1.py
@@ -1,17 +1,11 @@
-#!/usr/bin/env python
-import json_numpy as json
 import scipy.sparse
-import sys
-import glob
-import pyvista as pv
 import os
 
-import meshio
 import numpy as np
 from tqdm import tqdm
 
 from spice_util import get_sunvec
-from flux.compressed_form_factors_nmf import CompressedFormFactorMatrix
+from flux.compressed_form_factors import CompressedFormFactorMatrix
 from flux.model import ThermalModel
 from flux.shape import CgalTrimeshShapeModel, get_surface_normals
 
@@ -21,31 +15,26 @@ import argparse
 import arrow
 
 parser = argparse.ArgumentParser()
-parser.add_argument('--compression_type', type=str, default="svd",choices=["nmf","snmf","wsnmf",
-    "svd","ssvd",
-    "rand_svd","rand_ssvd","rand_snmf",
-    "saca","sbrp","rand_sid",
-    "stoch_radiosity",
-    "true_model"])
+parser.add_argument('--compression_type', type=str, default="svd", choices=["svd", "rand_svd", "aca", "rand_id", "paige", "sparse_tol", "sparse_k", "true_model"])
 parser.add_argument('--max_area', type=float, default=3.0)
 parser.add_argument('--outer_radius', type=int, default=80)
-parser.add_argument('--tol', type=float, default=1e-1)
 
+parser.add_argument('--tol', type=float, default=1e-1)
 parser.add_argument('--min_depth', type=int, default=1)
 parser.add_argument('--max_depth', type=int, default=0)
+parser.add_argument('--compress_sparse', action='store_true')
 
-parser.add_argument('--nmf_max_iters', type=int, default=int(1e4))
-parser.add_argument('--nmf_tol', type=float, default=1e-2)
-
+parser.add_argument('--add_residuals', action='store_true')
 parser.add_argument('--k0', type=int, default=40)
-
 parser.add_argument('--p', type=int, default=5)
 parser.add_argument('--q', type=int, default=1)
 
-parser.add_argument('--nmf_beta_loss', type=int, default=2, choices=[1,2])
+parser.add_argument('--paige_mult', type=int, default=1)
+parser.add_argument('--sparse_mult', type=int, default=1)
 
 parser.add_argument('--cliques', action='store_true')
 parser.add_argument('--n_cliques', type=int, default=25)
+parser.add_argument('--obb', action='store_true')
 
 parser.set_defaults(feature=False)
 
@@ -63,67 +52,53 @@ max_depth = args.max_depth if args.max_depth != 0 else None
 if compression_type == "true_model":
     FF_dir = "true_{}_{}".format(max_area_str, outer_radius_str)
 
-elif compression_type == "stoch_radiosity":
-    FF_dir = "stoch_rad_{}_{}_{}k0".format(max_area_str, outer_radius_str,
+elif compression_type == "paige":
+    FF_dir = "paige_{}_{}_{}k".format(max_area_str, outer_radius_str,
+        args.paige_mult)
+
+elif compression_type == "sparse_tol":
+    FF_dir = "sparse_{}_{}_{:.0e}".format(max_area_str, outer_radius_str,
+        args.tol)
+
+elif compression_type == "sparse_k":
+    FF_dir = "sparse_{}_{}_{}k".format(max_area_str, outer_radius_str,
+        args.sparse_mult)
+
+elif compression_type == "svd" or compression_type == "aca":
+    if args.add_residuals:
+        FF_dir = "{}_resid_{}_{}_{:.0e}_{}k0".format(compression_type, max_area_str, outer_radius_str, args.tol,
         args.k0)
+    else:
+        FF_dir = "{}_{}_{}_{:.0e}_{}k0".format(compression_type, max_area_str, outer_radius_str, args.tol,
+            args.k0)
 
-elif compression_type == "svd":
-    FF_dir = "{}_{}_{}_{:.0e}_{}k0".format(compression_type, max_area_str, outer_radius_str, args.tol,
-        args.k0)
-
-elif compression_type == "ssvd":
-    FF_dir = "{}_{}_{}_{:.0e}_{}k0".format(compression_type, max_area_str, outer_radius_str, args.tol,
-        args.k0)
-
-elif compression_type == "rand_svd":
-    FF_dir = "{}_{}_{}_{:.0e}_{}p_{}q_{}k0".format(compression_type, max_area_str, outer_radius_str, args.tol,
-        args.p, args.q, args.k0)
-
-elif compression_type == "rand_ssvd":
-    FF_dir = "{}_{}_{}_{:.0e}_{}p_{}q_{}k0".format(compression_type, max_area_str, outer_radius_str, args.tol,
-        args.p, args.q, args.k0)
-
-elif compression_type == "nmf":
-    FF_dir = "{}_{}_{}_{:.0e}_{:.0e}it_{:.0e}tol_{}k0".format(compression_type if args.nmf_beta_loss==2 else "klnmf", max_area_str, outer_radius_str, args.tol,
-        args.nmf_max_iters, args.nmf_tol, args.k0)
-
-elif compression_type == "snmf":
-    FF_dir = "{}_{}_{}_{:.0e}_{:.0e}it_{:.0e}tol_{}k0".format(compression_type if args.nmf_beta_loss==2 else "sklnmf", max_area_str, outer_radius_str, args.tol,
-        args.nmf_max_iters, args.nmf_tol, args.k0)
-
-elif compression_type == "rand_snmf":
-    FF_dir = "{}_{}_{}_{:.0e}_{:.0e}it_{:.0e}tol_{}p_{}q_{}k0".format(compression_type, max_area_str, outer_radius_str, args.tol,
-        args.nmf_max_iters, args.nmf_tol, args.p, args.q, args.k0)
-
-elif compression_type == "wsnmf":
-    FF_dir = "{}_{}_{}_{:.0e}_{:.0e}it_{:.0e}tol_{}k0".format(compression_type if args.nmf_beta_loss==2 else "wsklnmf", max_area_str, outer_radius_str, args.tol,
-        args.nmf_max_iters, args.nmf_tol, args.k0)
-
-elif compression_type == "saca":
-    FF_dir = "{}_{}_{}_{:.0e}_{}k0".format(compression_type, max_area_str, outer_radius_str, args.tol,
-        args.k0)
-
-elif compression_type == "sbrp":
-    FF_dir = "{}_{}_{}_{:.0e}_{}k0".format(compression_type, max_area_str, outer_radius_str, args.tol,
-        args.k0)
-
-elif compression_type == "rand_sid":
-    FF_dir = "{}_{}_{}_{:.0e}_{}p_{}q_{}k0".format(compression_type, max_area_str, outer_radius_str, args.tol,
-        args.p, args.q, args.k0)
+elif compression_type == "rand_svd" or compression_type == "rand_id":
+    if args.add_residuals:
+        FF_dir = "{}_resid_{}_{}_{:.0e}_{}p_{}q_{}k0".format(compression_type, max_area_str, outer_radius_str, args.tol,
+            args.p, args.q, args.k0)
+    else:
+        FF_dir = "{}_{}_{}_{:.0e}_{}p_{}q_{}k0".format(compression_type, max_area_str, outer_radius_str, args.tol,
+            args.p, args.q, args.k0)
 
 
-if not (compression_type == "true_model" or compression_type == "stoch_radiosity") and args.min_depth != 1:
+if not (compression_type == "true_model" or compression_type == "paige" or compression_type == "sparse_tol" or compression_type == "sparse_k") and args.min_depth != 1:
     FF_dir += "_{}mindepth".format(args.min_depth)
 
-if not (compression_type == "true_model" or compression_type == "stoch_radiosity") and max_depth is not None:
+if not (compression_type == "true_model" or compression_type == "paige" or compression_type == "sparse_tol" or compression_type == "sparse_k") and max_depth is not None:
     FF_dir += "_{}maxdepth".format(max_depth)
 
+if not (compression_type == "true_model" or compression_type == "paige" or compression_type == "sparse_tol" or compression_type == "sparse_k") and args.compress_sparse:
+    FF_dir += "_cs"
 
-if args.cliques:
+if not (compression_type == "true_model" or compression_type == "paige" or compression_type == "sparse_tol" or compression_type == "sparse_k") and args.cliques:
     FF_dir = "results_cliques/"+FF_dir
     FF_dir = FF_dir+"_{}nc".format(args.n_cliques)
 else:
     FF_dir = "results/"+FF_dir
+
+if not (compression_type == "true_model" or compression_type == "paige" or compression_type == "sparse_tol" or compression_type == "sparse_k") and args.cliques and args.obb:
+    FF_dir = FF_dir + "_obb"
+
 if not os.path.exists(FF_dir):
     print("PATH DOES NOT EXIST "+FF_dir)
     assert False
@@ -131,7 +106,7 @@ savedir = FF_dir
 
 
 # read shapemodel and form-factor matrix generated by make_compressed_form_factor_matrix.py
-if compression_type == 'true_model' or compression_type == 'stoch_radiosity':
+if compression_type == 'true_model' or compression_type == "paige" or compression_type == "sparse_tol" or compression_type == "sparse_k":
     path = FF_dir+f'/FF_{max_area_str}_{outer_radius_str}.npz'
     FF = scipy.sparse.load_npz(path)
     V = np.load(f'shackleton_verts_{max_area_str}_{outer_radius_str}.npy')
@@ -140,7 +115,10 @@ if compression_type == 'true_model' or compression_type == 'stoch_radiosity':
     N[N[:, 2] > 0] *= -1
     shape_model = CgalTrimeshShapeModel(V, F, N)
 else:
-    path = FF_dir+f'/FF_{max_area_str}_{outer_radius_str}_{tol_str}_{compression_type}.bin'
+    if args.add_residuals:
+        path = FF_dir+f'/FF_{max_area_str}_{outer_radius_str}_{tol_str}_{compression_type}_resid.bin'
+    else:
+        path = FF_dir+f'/FF_{max_area_str}_{outer_radius_str}_{tol_str}_{compression_type}.bin'
     FF = CompressedFormFactorMatrix.from_file(path)
     shape_model = FF.shape_model
