+++ examples/shackleton_vary_outer/make_block_plots.py
@@ -3,20 +3,33 @@ import matplotlib.patches as patches
 import matplotlib.pyplot as plt
 import numpy as np
 
-import flux.compressed_form_factors_truncated_sparse as cff
-from flux.compressed_form_factors_truncated_sparse import TruncatedHierarchicalFormFactorMatrix
+import flux.compressed_form_factors as cff
+from flux.compressed_form_factors import CompressedFormFactorMatrix
 
 import scipy
 import os
 import argparse
 
 parser = argparse.ArgumentParser()
+parser.add_argument('--compression_type', type=str, default="svd", choices=["svd", "rand_svd", "aca", "rand_id"])
 parser.add_argument('--max_area', type=float, default=3.0)
 parser.add_argument('--outer_radius', type=int, default=80)
-parser.add_argument('--tol', type=float, default=1e-1)
 
+parser.add_argument('--tol', type=float, default=1e-1)
 parser.add_argument('--min_depth', type=int, default=1)
 parser.add_argument('--max_depth', type=int, default=0)
+parser.add_argument('--compress_sparse', action='store_true')
+
+parser.add_argument('--add_residuals', action='store_true')
+parser.add_argument('--k0', type=int, default=40)
+parser.add_argument('--p', type=int, default=5)
+parser.add_argument('--q', type=int, default=1)
+
+parser.add_argument('--cliques', action='store_true')
+parser.add_argument('--n_cliques', type=int, default=25)
+parser.add_argument('--obb', action='store_true')
+
+parser.add_argument('--plot_labels', action='store_true')
 
 parser.set_defaults(feature=False)
 
@@ -57,7 +70,9 @@ def plot_blocks(block, fig, **kwargs):
 
             child = block._blocks[i, j]
             if child.is_leaf:
-                if isinstance(child, cff.FormFactorTruncatedSparseBlock):
+                if isinstance(child, cff.FormFactorSvdBlock) or isinstance(child, cff.FormFactorSparseSvdBlock) or \
+                isinstance(child, cff.FormFactorAcaBlock) or isinstance(child, cff.FormFactorSparseAcaBlock) or \
+                isinstance(child, cff.FormFactorIdBlock) or isinstance(child, cff.FormFactorSparseIdBlock):
                     facecolor = 'cyan' if child.compressed else 'orange'
                     rect = patches.Rectangle(
                         c, w, h, edgecolor='none', facecolor=facecolor)
@@ -78,6 +93,9 @@ def plot_blocks(block, fig, **kwargs):
                     continue
                 else:
                     raise Exception('TODO: add %s to _plot_block' % type(child))
+
+                if args.plot_labels:
+                    ax.text(c[0] + 0.5*w, c[1] + 0.5*h, "{:.0e}".format(np.prod(child.shape)), transform=ax.transAxes, fontsize=8, verticalalignment='center', horizontalalignment='center')
             else:
                 add_rects(child, c, w, h)
 
@@ -94,30 +112,62 @@ def plot_blocks(block, fig, **kwargs):
 
 
 
+compression_type = args.compression_type
 max_area_str = str(args.max_area)
 outer_radius_str = str(args.outer_radius)
 
 max_depth = args.max_depth if args.max_depth != 0 else None
 
-FF_dir = "sparse_hier_{}_{}_{:.0e}".format(max_area_str, outer_radius_str, args.tol)
+if compression_type == "svd" or compression_type == "aca":
+    if args.add_residuals:
+        FF_dir = "{}_resid_{}_{}_{:.0e}_{}k0".format(compression_type, max_area_str, outer_radius_str, args.tol,
+        args.k0)
+    else:
+        FF_dir = "{}_{}_{}_{:.0e}_{}k0".format(compression_type, max_area_str, outer_radius_str, args.tol,
+            args.k0)
+
+elif compression_type == "rand_svd" or compression_type == "rand_id":
+    if args.add_residuals:
+        FF_dir = "{}_resid_{}_{}_{:.0e}_{}p_{}q_{}k0".format(compression_type, max_area_str, outer_radius_str, args.tol,
+            args.p, args.q, args.k0)
+    else:
+        FF_dir = "{}_{}_{}_{:.0e}_{}p_{}q_{}k0".format(compression_type, max_area_str, outer_radius_str, args.tol,
+            args.p, args.q, args.k0)
 
 
-if args.min_depth != 1:
+if not args.cliques and args.min_depth != 1:
     FF_dir += "_{}mindepth".format(args.min_depth)
 
 if max_depth is not None:
     FF_dir += "_{}maxdepth".format(max_depth)
 
+if args.compress_sparse:
+    FF_dir += "_cs"
 
-FF_path =  "results/"+ FF_dir + "/FF_{}_{}_{:.0e}_sparse_hierarch.bin".format(max_area_str, outer_radius_str, args.tol)
+result_dir = "results"
+if args.cliques:
+    FF_dir += "_{}nc".format(args.n_cliques)
+    result_dir += "_cliques"
+
+if args.cliques and args.obb:
+    FF_dir = FF_dir + "_obb"
+
+if args.add_residuals:
+    FF_path =  result_dir+"/"+ FF_dir + "/FF_{}_{}_{:.0e}_{}_resid.bin".format(max_area_str, outer_radius_str, args.tol, compression_type)
+else:
+    FF_path =  result_dir+"/"+ FF_dir + "/FF_{}_{}_{:.0e}_{}.bin".format(max_area_str, outer_radius_str, args.tol, compression_type)
+
 if not os.path.exists(FF_path):
     print("PATH DOES NOT EXIST " + FF_path)
     assert False
-FF = TruncatedHierarchicalFormFactorMatrix.from_file(FF_path)
+FF = CompressedFormFactorMatrix.from_file(FF_path)
 
 
 fig = plt.figure(figsize=(18, 6))
 print(f'- {FF_path}')
 fig, ax = plot_blocks(FF._root, fig)
-fig.savefig("results/"+FF_dir+"/block_plot.png")
+if args.plot_labels:
+    fig.savefig(result_dir+"/"+FF_dir+"/block_plot_labeled.png")
+else:
+    fig.savefig(result_dir+"/"+FF_dir+"/block_plot.png")
 plt.close(fig)
