+++ src/flux/compressed_form_factors.py
@@ -18,7 +18,8 @@ from flux.debug import IndentedPrinter
 from flux.form_factors import get_form_factor_matrix
 from flux.octree import get_octant_order
 from flux.quadtree import get_quadrant_order
-from flux.util import nbytes
+from flux.obbtree import get_obb_partition_2d
+from flux.util import nbytes, get_sunvec
 
 
 @np.vectorize
@@ -90,12 +91,17 @@ class FormFactorNullBlock(FormFactorLeafBlock,
             raise RuntimeError('a null block must have a degenerate shape')
         super().__init__(root, shape)
 
+        self.partial_norm = 0.
+
     def _matmat(self, x):
         if self.shape[1] != x.shape[0]:
             raise ValueError(
                 'multiplying null block with matrix with wrong shape')
         new_shape = (self.shape[0], x.shape[1])
         return np.empty(new_shape, dtype=self.dtype)
+
+    def __add__(self, x):
+        return np.empty(self.shape, dtype=self.dtype)
 
     def is_dense(self):
         return True
@@ -115,16 +121,25 @@ class FormFactorNullBlock(FormFactorLeafBlock,
     def is_empty_leaf(self):
         return True
 
+    @property
+    def sq_resid_sum(self):
+        return 0.
+
 
 class FormFactorZeroBlock(FormFactorLeafBlock,
                           scipy.sparse.linalg.LinearOperator):
     def __init__(self, root, shape):
         super().__init__(root, shape)
 
+        self.partial_norm = 0.
+
     def _matmat(self, x):
         m = self.shape[0]
         y_shape = (m,) if x.ndim == 1 else (m, x.shape[1])
         return np.zeros(y_shape, dtype=self.dtype)
+
+    def __add__(self, x):
+        return x
 
     def is_dense(self):
         return False
@@ -147,6 +162,10 @@ class FormFactorZeroBlock(FormFactorLeafBlock,
     def is_empty_leaf(self):
         return True
 
+    @property
+    def sq_resid_sum(self):
+        return 0.
+
 
 class FormFactorDenseBlock(FormFactorLeafBlock,
                            scipy.sparse.linalg.LinearOperator):
@@ -155,6 +174,8 @@ class FormFactorDenseBlock(FormFactorLeafBlock,
             mat = mat.toarray()
         super().__init__(root, mat.shape)
         self._mat = mat
+
+        self.partial_norm = np.power(mat.flatten(), 2).sum()
 
     @property
     def nbytes(self):
@@ -175,6 +196,9 @@ class FormFactorDenseBlock(FormFactorLeafBlock,
     def _matmat(self, x):
         return self._mat@x
 
+    def __add__(self, x):
+        return self._mat+x
+
     def _get_sparsity(self, tol=None):
         nnz = flux.linalg.nnz(self._mat, tol)
         size = self._mat.size
@@ -184,6 +208,10 @@ class FormFactorDenseBlock(FormFactorLeafBlock,
     def is_empty_leaf(self):
         return False
 
+    @property
+    def sq_resid_sum(self):
+        return 0.
+
 
 class FormFactorSparseBlock(FormFactorLeafBlock,
                             scipy.sparse.linalg.LinearOperator):
@@ -192,6 +220,9 @@ class FormFactorSparseBlock(FormFactorLeafBlock,
 
     def _matmat(self, x):
         return np.array(self._spmat@x)
+
+    def __add__(self, x):
+        return np.array(self._spmat+x)
 
     def is_dense(self):
         return False
@@ -203,14 +234,20 @@ class FormFactorSparseBlock(FormFactorLeafBlock,
     def is_empty_leaf(self):
         return False
 
+    @property
+    def sq_resid_sum(self):
+        return 0.
+
 
 class FormFactorCsrBlock(FormFactorSparseBlock):
 
     def __init__(self, root, mat):
         super().__init__(root, mat.shape)
         if isinstance(mat, np.ndarray):
+            self.partial_norm = np.power(mat.flatten(), 2).sum()
             spmat = scipy.sparse.csr_matrix(mat)
         elif isinstance(mat, scipy.sparse.spmatrix):
+            self.partial_norm = np.power(mat.A.flatten(), 2).sum()
             spmat = mat
         else:
             raise Exception('invalid class for mat: %s' % type(mat))
@@ -227,11 +264,66 @@ class FormFactorCsrBlock(FormFactorSparseBlock):
     def _mat(self):
         return self._spmat.toarray()
 
+    @property
+    def sq_resid_sum(self):
+        return 0.
+
+
+class FormFactorTruncatedCsrBlock(FormFactorSparseBlock):
+
+    def __init__(self, root, mat):
+        super().__init__(root, mat.shape)
+        if isinstance(mat, np.ndarray):
+            dmat = mat
+        elif isinstance(mat, scipy.sparse.spmatrix):
+            dmat = mat.A
+        else:
+            raise Exception('invalid class for mat: %s' % type(mat))
+        
+        Sr = self.get_form_factor_truncated(dmat)
+        self._spmat = Sr
+        
+        self.partial_norm = np.power(dmat.flatten(), 2).sum()
+        self._sq_resid_sum = np.power((self._spmat - dmat).flatten(), 2).sum()
+
+    def get_form_factor_truncated(self, mat):
+
+        num_elements = mat.shape[0] * mat.shape[1]
+
+        sorted_ff_idx = np.unravel_index(np.argsort(abs(mat), axis=None), mat.shape)
+
+        target = np.power(np.linalg.norm(mat, ord='fro'), 2) - np.power(0.1*self._tol*np.linalg.norm(mat, ord='fro'), 2)
+
+        cumulative_residual = np.cumsum(np.power(mat[sorted_ff_idx[0], sorted_ff_idx[1]][::-1], 2))
+        keep_resids = (cumulative_residual > target).nonzero()[0][0] + 1
+
+        Sr = np.copy(mat)
+        Sr[sorted_ff_idx[0][:num_elements-keep_resids], sorted_ff_idx[1][:num_elements-keep_resids]] = 0.
+
+        Sr = scipy.sparse.csr_matrix(Sr)
+
+        return Sr
+
+    @property
+    def nbytes(self):
+        return self._spmat.data.nbytes
+
+    def tocsr(self):
+        return self._spmat
+
+    @property
+    def _mat(self):
+        return self._spmat.toarray()
+
+    @property
+    def sq_resid_sum(self):
+        return self._sq_resid_sum
+
 
 class FormFactorSvdBlock(FormFactorLeafBlock,
                          scipy.sparse.linalg.LinearOperator):
 
-    def __init__(self, root, u, s, vt):
+    def __init__(self, root, u, s, vt, mat_block):
         shape = (u.shape[0], vt.shape[1])
         super().__init__(root, shape)
 
@@ -244,11 +336,17 @@ class FormFactorSvdBlock(FormFactorLeafBlock,
         vt_csr = scipy.sparse.csr_matrix(vt)
         self._vt = vt if nbytes(vt) < nbytes(vt_csr) else vt_csr
 
+        self.partial_norm = np.power(mat_block.flatten(), 2).sum()
+        self.sq_resid_sum = np.power(((self._u @ np.diag(self._s) @ self._vt) - mat_block).flatten(), 2).sum()
+
     def _matmat(self, x):
         y = self._vt@x
         y = (y.T*self._s).T
         y = self._u@y
         return y
+
+    def __add__(self, x):
+        return (self._u @ np.diag(self._s) @ self._vt) + x
 
     def is_dense(self):
         return False
@@ -276,6 +374,278 @@ class FormFactorSvdBlock(FormFactorLeafBlock,
         return 0 if size == 0 else (u_nnz + v_nnz + self._k)/size
 
 
+class FormFactorSparseSvdBlock(FormFactorLeafBlock,
+                         scipy.sparse.linalg.LinearOperator):
+
+    def __init__(self, root, u, s, vt, sr, mat_block):
+        shape = (u.shape[0], vt.shape[1])
+        super().__init__(root, shape)
+
+        self._k = s.size
+        self._s = s
+
+        u_csr = scipy.sparse.csr_matrix(u)
+        self._u = u if nbytes(u) < nbytes(u_csr) else u_csr
+
+        vt_csr = scipy.sparse.csr_matrix(vt)
+        self._vt = vt if nbytes(vt) < nbytes(vt_csr) else vt_csr
+
+        sr_csr = scipy.sparse.csr_matrix(sr)
+        self._sr = sr if nbytes(sr) < nbytes(sr_csr) else sr_csr
+
+        self.partial_norm = np.power(mat_block.flatten(), 2).sum()
+        self.sq_resid_sum = np.power(((self._u @ np.diag(self._s) @ self._vt) + self._sr - mat_block).flatten(), 2).sum()
+
+    def _matmat(self, x):
+        y = self._vt@x
+        y = (y.T*self._s).T
+        y = self._u@y
+        y = y + (self._sr@x)
+        return y
+
+    def __add__(self, x):
+        return (self._u @ np.diag(self._s) @ self._vt) + self._sr + x
+
+    def is_dense(self):
+        return False
+
+    def is_sparse(self):
+        return False
+
+    @property
+    def is_empty_leaf(self):
+        return False
+
+    @property
+    def nbytes(self):
+        return nbytes(self._u) + nbytes(self._s) + nbytes(self._vt) + nbytes(self._sr)
+
+    @property
+    def compressed(self):
+        return isinstance(self._u, scipy.sparse.spmatrix) \
+            or isinstance(self._vt, scipy.sparse.spmatrix) \
+            or isinstance(self._sr, scipy.sparse.spmatrix)
+
+    def _get_sparsity(self, tol=None):
+        u_nnz = flux.linalg.nnz(self._u, tol)
+        v_nnz = flux.linalg.nnz(self._vt, tol)
+        sr_nnz = flux.linalg.nnz(self._sr, tol)
+        size = self._u.size + self._vt.size + self._k + self._sr.size
+        return 0 if size == 0 else (u_nnz + v_nnz + self._k + sr_nnz)/size
+
+
+class FormFactorAcaBlock(FormFactorLeafBlock,
+                         scipy.sparse.linalg.LinearOperator):
+
+    def __init__(self, root, a, b, mat_block):
+        shape = (a.shape[0], b.shape[1])
+        super().__init__(root, shape)
+
+        a_csr = scipy.sparse.csr_matrix(a)
+        self._a = a if nbytes(a) < nbytes(a_csr) else a_csr
+
+        b_csr = scipy.sparse.csr_matrix(b)
+        self._b = b if nbytes(b) < nbytes(b_csr) else b_csr
+
+        self.partial_norm = np.power(mat_block.flatten(), 2).sum()
+        self.sq_resid_sum = np.power((self._a@self._b - mat_block).flatten(), 2).sum()
+
+    def _matmat(self, x):
+        y = self._b@x
+        y = self._a@y
+        return y
+
+    def __add__(self, x):
+        return self._a@self._b + x
+
+    def is_dense(self):
+        return False
+
+    def is_sparse(self):
+        return False
+
+    @property
+    def is_empty_leaf(self):
+        return False
+
+    @property
+    def nbytes(self):
+        return nbytes(self._a) + nbytes(self._b)
+
+    @property
+    def compressed(self):
+        return isinstance(self._a, scipy.sparse.spmatrix) \
+            or isinstance(self._b, scipy.sparse.spmatrix)
+
+    def _get_sparsity(self, tol=None):
+        a_nnz = flux.linalg.nnz(self._a, tol)
+        b_nnz = flux.linalg.nnz(self._b, tol)
+        size = self._a.size + self._b.size
+        return 0 if size == 0 else (a_nnz + b_nnz)/size
+
+
+class FormFactorSparseAcaBlock(FormFactorLeafBlock,
+                         scipy.sparse.linalg.LinearOperator):
+
+    def __init__(self, root, a, b, sr, mat_block):
+        shape = (a.shape[0], b.shape[1])
+        super().__init__(root, shape)
+
+        a_csr = scipy.sparse.csr_matrix(a)
+        self._a = a if nbytes(a) < nbytes(a_csr) else a_csr
+
+        b_csr = scipy.sparse.csr_matrix(b)
+        self._b = b if nbytes(b) < nbytes(b_csr) else b_csr
+
+        sr_csr = scipy.sparse.csr_matrix(sr)
+        self._sr = sr if nbytes(sr) < nbytes(sr_csr) else sr_csr
+
+        self.partial_norm = np.power(mat_block.flatten(), 2).sum()
+        self.sq_resid_sum = np.power((self._a@self._b + self._sr - mat_block).flatten(), 2).sum()
+
+    def _matmat(self, x):
+        y = self._b@x
+        y = self._a@y
+        y = y + (self._sr@x)
+        return y
+
+    def __add__(self, x):
+        return self._a@self._b + self._sr + x
+
+    def is_dense(self):
+        return False
+
+    def is_sparse(self):
+        return False
+
+    @property
+    def is_empty_leaf(self):
+        return False
+
+    @property
+    def nbytes(self):
+        return nbytes(self._a) + nbytes(self._b) + nbytes(self._sr)
+
+    @property
+    def compressed(self):
+        return isinstance(self._a, scipy.sparse.spmatrix) \
+            or isinstance(self._b, scipy.sparse.spmatrix) \
+            or isinstance(self._sr, scipy.sparse.spmatrix)
+
+    def _get_sparsity(self, tol=None):
+        a_nnz = flux.linalg.nnz(self._a, tol)
+        b_nnz = flux.linalg.nnz(self._b, tol)
+        sr_nnz = flux.linalg.nnz(self._sr, tol)
+        size = self._a.size + self._b.size + self._sr.size
+        return 0 if size == 0 else (a_nnz + b_nnz + sr_nnz)/size
+
+
+class FormFactorIdBlock(FormFactorLeafBlock,
+                         scipy.sparse.linalg.LinearOperator):
+
+    def __init__(self, root, c, v, mat_block):
+        shape = (c.shape[0], v.shape[1])
+        super().__init__(root, shape)
+
+        c_csr = scipy.sparse.csr_matrix(c)
+        self._c = c if nbytes(c) < nbytes(c_csr) else c_csr
+
+        v_csr = scipy.sparse.csr_matrix(v)
+        self._v = v if nbytes(v) < nbytes(v_csr) else v_csr
+
+        self.partial_norm = np.power(mat_block.flatten(), 2).sum()
+        self.sq_resid_sum = np.power((self._c@self._v - mat_block).flatten(), 2).sum()
+
+    def _matmat(self, x):
+        y = self._v@x
+        y = self._c@y
+        return y
+
+    def __add__(self, x):
+        return self._c@self._v + x
+
+    def is_dense(self):
+        return False
+
+    def is_sparse(self):
+        return False
+
+    @property
+    def is_empty_leaf(self):
+        return False
+
+    @property
+    def nbytes(self):
+        return nbytes(self._c) + nbytes(self._v)
+
+    @property
+    def compressed(self):
+        return isinstance(self._c, scipy.sparse.spmatrix) \
+            or isinstance(self._v, scipy.sparse.spmatrix)
+
+    def _get_sparsity(self, tol=None):
+        c_nnz = flux.linalg.nnz(self._c, tol)
+        v_nnz = flux.linalg.nnz(self._v, tol)
+        size = self._c.size + self._v.size
+        return 0 if size == 0 else (c_nnz + v_nnz)/size
+
+
+class FormFactorSparseIdBlock(FormFactorLeafBlock,
+                         scipy.sparse.linalg.LinearOperator):
+
+    def __init__(self, root, c, v, sr, mat_block):
+        shape = (c.shape[0], v.shape[1])
+        super().__init__(root, shape)
+
+        c_csr = scipy.sparse.csr_matrix(c)
+        self._c = c if nbytes(c) < nbytes(c_csr) else c_csr
+
+        v_csr = scipy.sparse.csr_matrix(v)
+        self._v = v if nbytes(v) < nbytes(v_csr) else v_csr
+
+        sr_csr = scipy.sparse.csr_matrix(sr)
+        self._sr = sr if nbytes(sr) < nbytes(sr_csr) else sr_csr
+
+        self.partial_norm = np.power(mat_block.flatten(), 2).sum()
+        self.sq_resid_sum = np.power((self._c@self._v + self._sr - mat_block).flatten(), 2).sum()
+
+    def _matmat(self, x):
+        y = self._v@x
+        y = self._c@y
+        y = y + (self._sr@x)
+        return y
+
+    def __add__(self, x):
+        return self._c@self._v + self._sr + x
+
+    def is_dense(self):
+        return False
+
+    def is_sparse(self):
+        return False
+
+    @property
+    def is_empty_leaf(self):
+        return False
+
+    @property
+    def nbytes(self):
+        return nbytes(self._c) + nbytes(self._v) + nbytes(self._sr)
+
+    @property
+    def compressed(self):
+        return isinstance(self._c, scipy.sparse.spmatrix) \
+            or isinstance(self._v, scipy.sparse.spmatrix) \
+            or isinstance(self._sr, scipy.sparse.spmatrix)
+
+    def _get_sparsity(self, tol=None):
+        c_nnz = flux.linalg.nnz(self._c, tol)
+        v_nnz = flux.linalg.nnz(self._v, tol)
+        sr_nnz = flux.linalg.nnz(self._sr, tol)
+        size = self._c.size + self._v.size + self._sr.size
+        return 0 if size == 0 else (c_nnz + v_nnz + sr_nnz)/size
+
+
 class FormFactorBlockMatrix(CompressedFormFactorBlock,
                             scipy.sparse.linalg.LinearOperator):
 
@@ -283,7 +653,7 @@ class FormFactorBlockMatrix(CompressedFormFactorBlock,
         super().__init__(root, shape)
 
     def make_block(self, shape_model, I, J, spmat,
-                   max_depth=None, force_max_depth=False):
+                   max_depth=None, force_max_depth=False, compression_type="svd", compression_params={}):
         if max_depth is not None and not isinstance(max_depth, int):
             raise RuntimeError(
                 'invalid max_depth type: %s' % str(type(max_depth)))
@@ -340,10 +710,91 @@ class FormFactorBlockMatrix(CompressedFormFactorBlock,
             else:
                 return sparse_block
 
+
         # Next, since the block is "big enough", we go ahead and
         # attempt to compress it.
-        svd_block = self._get_svd_block(spmat)
-        nbytes_svd = np.inf if svd_block is None else nbytes(svd_block)
+
+        if compression_type == "svd":
+
+            if not 'k0' in compression_params.keys():
+                raise RuntimeError('compression_params must include key: k0')
+            
+            k0 = compression_params["k0"]
+            if self.root._add_residuals:
+                compressed_block = self._get_sparse_svd_block(spmat, k0=k0)
+                nbytes_compressed = np.inf if compressed_block is None else nbytes(compressed_block)
+            else:
+                compressed_block = self._get_svd_block(spmat, k0=k0)
+                nbytes_compressed = np.inf if compressed_block is None else nbytes(compressed_block)
+
+
+        elif compression_type == "rand_svd":
+
+            if not np.array([key in compression_params.keys() for key in ['k0', 'p', 'q']]).all():
+                raise RuntimeError('compression_params must include keys: %s' % ['k0', 'p', 'q'])
+            
+            k0 = compression_params["k0"]
+            p = compression_params["p"]
+            q = compression_params["q"]
+
+            if self.root._add_residuals:
+                compressed_block = self._get_sparse_random_svd_block(spmat, k0=k0, p=p, q=q)
+                nbytes_compressed = np.inf if compressed_block is None else nbytes(compressed_block)
+            else:
+                compressed_block = self._get_random_svd_block(spmat, k0=k0, p=p, q=q)
+                nbytes_compressed = np.inf if compressed_block is None else nbytes(compressed_block)
+
+
+        elif compression_type == "aca":
+
+            if not 'k0' in compression_params.keys():
+                raise RuntimeError('compression_params must include key: k0')
+            
+            k0 = compression_params["k0"]
+
+            if self.root._add_residuals:
+                compressed_block = self._get_sparse_aca_block(spmat, k0=k0)
+                nbytes_compressed = np.inf if compressed_block is None else nbytes(compressed_block)
+            else:
+                compressed_block = self._get_aca_block(spmat, k0=k0)
+                nbytes_compressed = np.inf if compressed_block is None else nbytes(compressed_block)
+
+
+        elif compression_type == "paca":
+
+            if not 'k0' in compression_params.keys():
+                raise RuntimeError('compression_params must include key: k0')
+            
+            k0 = compression_params["k0"]
+
+            if self.root._add_residuals:
+                compressed_block = self._get_sparse_paca_block(spmat, k0=k0)
+                nbytes_compressed = np.inf if compressed_block is None else nbytes(compressed_block)
+            else:
+                compressed_block = self._get_paca_block(spmat, k0=k0)
+                nbytes_compressed = np.inf if compressed_block is None else nbytes(compressed_block)
+
+
+        elif compression_type == "rand_id":
+
+            if not np.array([key in compression_params.keys() for key in ['k0', 'p', 'q']]).all():
+                raise RuntimeError('compression_params must include keys: %s' % ['k0', 'p', 'q'])
+            
+            k0 = compression_params["k0"]
+            p = compression_params["p"]
+            q = compression_params["q"]
+
+            if self.root._add_residuals:
+                compressed_block = self._get_sparse_random_id_block(spmat, k0=k0, p=p, q=q)
+                nbytes_compressed = np.inf if compressed_block is None else nbytes(compressed_block)
+            else:
+                compressed_block = self._get_random_id_block(spmat, k0=k0, p=p, q=q)
+                nbytes_compressed = np.inf if compressed_block is None else nbytes(compressed_block)
+
+
+        else:
+            raise RuntimeError('invalid compression_type: %d' % compression_type)
+
 
         # If we haven't specified a max depth, or if we haven't
         # bottomed out yet, then we attempt to descend another
@@ -356,15 +807,16 @@ class FormFactorBlockMatrix(CompressedFormFactorBlock,
             child_block = None
         nbytes_child = np.inf if child_block is None else nbytes(child_block)
 
-        nbytes_min = min(nbytes_sparse, nbytes_svd, nbytes_child)
+        nbytes_min = min(nbytes_sparse, nbytes_compressed, nbytes_child)
 
         # Select the block with the smallest size
         if nbytes_sparse == nbytes_min:
             block = sparse_block
-        elif nbytes_svd == nbytes_min:
-            block = svd_block
+        elif nbytes_compressed == nbytes_min:
+            block = compressed_block
         else:
             block = child_block
+
 
         # Finally, do a little post-processing: if all of the child
         # blocks are dense blocks, then collapse them into a single
@@ -382,17 +834,25 @@ class FormFactorBlockMatrix(CompressedFormFactorBlock,
         # Finally, we return whatever we have at this point. This
         # should either be an instance of ChildBlock or an SVD block.
         assert isinstance(block, type(self)) \
-            or isinstance(block, FormFactorSvdBlock)
+            or isinstance(block, FormFactorSvdBlock) \
+            or isinstance(block, FormFactorSparseSvdBlock) \
+            or isinstance(block, FormFactorAcaBlock) \
+            or isinstance(block, FormFactorSparseAcaBlock) \
+            or isinstance(block, FormFactorIdBlock) \
+            or isinstance(block, FormFactorSparseIdBlock) \
+            or isinstance(block, FormFactorQuadtreeBlock) \
+            or isinstance(block, FormFactorObbQuadtreeBlock)
         return block
 
-    def _get_svd_block(self, spmat):
+    def _get_svd_block(self, spmat, k0=40):
         ret = flux.linalg.estimate_rank(
-            spmat, self._tol, max_nbytes=nbytes(spmat))
+            spmat, self._tol, max_nbytes=nbytes(spmat),
+            k0=k0)
         if ret is None:
             return None
 
         U, S, Vt, tol = ret
-        svd_block = self.root.make_svd_block(U, S, Vt)
+        svd_block = self.root.make_svd_block(U, S, Vt, spmat.A)
 
         # If the tolerance estimated this way doesn't satisfy
         # the requested tolerance, return the sparse block
@@ -404,6 +864,146 @@ class FormFactorBlockMatrix(CompressedFormFactorBlock,
         a larger sparse block instead...""")
         return None
 
+    def _get_random_svd_block(self, spmat, k0=40, p=5, q=1):
+        ret = flux.linalg.estimate_rank_random_svd(
+            spmat, self._tol, max_nbytes=nbytes(spmat),
+            k0=k0, p=p, q=q)
+        if ret is None:
+            return None
+
+        U, S, Vt, tol = ret
+        svd_block = self.root.make_svd_block(U, S, Vt, spmat.A)
+
+        # If the tolerance estimated this way doesn't satisfy
+        # the requested tolerance, return the sparse block
+        # assert tol != 0
+        if tol <= self._tol:
+            return svd_block
+
+        logging.warning("""computed a really inaccurate SVD, using
+        a larger sparse block instead...""")
+        return None
+
+    def _get_sparse_svd_block(self, spmat, k0=40):
+        ret = flux.linalg.estimate_sparsity_svd(
+            spmat, self._tol, max_nbytes=nbytes(spmat),
+            k0=k0)
+        if ret is None:
+            return None
+
+        U, S, Vt, Sr = ret
+        s_svd_block = self.root.make_sparse_svd_block(U, S, Vt, Sr, spmat.A)
+
+        return s_svd_block
+
+    def _get_sparse_random_svd_block(self, spmat, k0=40, p=5, q=1):
+        ret = flux.linalg.estimate_sparsity_random_svd(
+            spmat, self._tol, max_nbytes=nbytes(spmat),
+            k0=k0, p=p, q=q)
+        if ret is None:
+            return None
+
+        U, S, Vt, Sr = ret
+        s_svd_block = self.root.make_sparse_svd_block(U, S, Vt, Sr, spmat.A)
+
+        return s_svd_block
+
+    def _get_aca_block(self, spmat, k0=40):
+        ret = flux.linalg.estimate_rank_aca(
+            spmat, self._tol, max_nbytes=nbytes(spmat),
+            k0=k0)
+        if ret is None:
+            return None
+
+        A, B, tol = ret
+        aca_block = self.root.make_aca_block(A, B, spmat.A)
+
+        # If the tolerance estimated this way doesn't satisfy
+        # the requested tolerance, return the sparse block
+        # assert tol != 0
+        if tol <= self._tol:
+            return aca_block
+
+        logging.warning("""computed a really inaccurate ACA, using
+        a larger sparse block instead...""")
+        return None
+
+    def _get_sparse_aca_block(self, spmat, k0=40):
+        ret = flux.linalg.estimate_sparsity_aca(
+            spmat, self._tol, max_nbytes=nbytes(spmat),
+            k0=k0)
+        if ret is None:
+            return None
+
+        A, B, Sr = ret
+        s_aca_block = self.root.make_sparse_aca_block(A, B, Sr, spmat.A)
+
+        return s_aca_block
+
+    def _get_paca_block(self, spmat, k0=40):
+        ret = flux.linalg.estimate_rank_partial_aca(
+            spmat, self._tol, max_nbytes=nbytes(spmat),
+            k0=k0)
+        if ret is None:
+            return None
+
+        A, B, tol = ret
+        aca_block = self.root.make_aca_block(A, B, spmat.A)
+
+        # If the tolerance estimated this way doesn't satisfy
+        # the requested tolerance, return the sparse block
+        # assert tol != 0
+        if tol <= self._tol:
+            return aca_block
+
+        logging.warning("""computed a really inaccurate ACA, using
+        a larger sparse block instead...""")
+        return None
+
+    def _get_sparse_paca_block(self, spmat, k0=40):
+        ret = flux.linalg.estimate_sparsity_partial_aca(
+            spmat, self._tol, max_nbytes=nbytes(spmat),
+            k0=k0)
+        if ret is None:
+            return None
+
+        A, B, Sr = ret
+        s_aca_block = self.root.make_sparse_aca_block(A, B, Sr, spmat.A)
+
+        return s_aca_block
+
+    def _get_random_id_block(self, spmat, k0=40, p=5, q=1):
+        ret = flux.linalg.estimate_rank_random_id(
+            spmat, self._tol, max_nbytes=nbytes(spmat),
+            k0=k0, p=p, q=q)
+        if ret is None:
+            return None
+
+        C, V, tol = ret
+        id_block = self.root.make_id_block(C, V, spmat.A)
+
+        # If the tolerance estimated this way doesn't satisfy
+        # the requested tolerance, return the sparse block
+        # assert tol != 0
+        if tol <= self._tol:
+            return id_block
+
+        logging.warning("""computed a really inaccurate ID, using
+        a larger sparse block instead...""")
+        return None
+
+    def _get_sparse_random_id_block(self, spmat, k0=40, p=5, q=1):
+        ret = flux.linalg.estimate_sparsity_random_id(
+            spmat, self._tol, max_nbytes=nbytes(spmat),
+            k0=k0, p=p, q=q)
+        if ret is None:
+            return None
+
+        C, V, Sr = ret
+        s_id_block = self.root.make_sparse_id_block(C, V, Sr, spmat.A)
+
+        return s_id_block
+
     def _matmat(self, x):
         y = np.zeros((self.shape[0], x.shape[1]), dtype=self.dtype)
         for i, row_inds in enumerate(self._row_block_inds):
@@ -412,6 +1012,37 @@ class FormFactorBlockMatrix(CompressedFormFactorBlock,
                 if block.is_empty_leaf: continue
                 y[row_inds] += block@x[col_inds]
         return y
+
+    def __add__(self, x):
+        if x.shape != self.shape:
+            raise ValueError('cannot add %r and %r shape object: shape mismatch' % (self, x.shape))
+        if isinstance(x, scipy.sparse.csr_matrix):
+            y = np.zeros((self.shape[0], self.shape[1]), dtype=self.dtype)
+        else:
+            y = np.zeros((self.shape[0], self.shape[1]), dtype=self.dtype)
+        for i, row_inds in enumerate(self._row_block_inds):
+            for j, col_inds in enumerate(self._col_block_inds):
+                block = self._blocks[i, j]
+                if block.is_empty_leaf: continue
+                update_inds = np.ix_(row_inds, col_inds)
+                y[update_inds] = block + x[update_inds]
+        return y
+
+    @property
+    def partial_norm(self):
+        total = 0.
+        for i in range(len(self._row_block_inds)):
+            for j in range(len(self._col_block_inds)):
+                total += self._blocks[i, j].partial_norm
+        return total
+
+    @property
+    def sq_resid_sum(self):
+        total = 0.
+        for i in range(len(self._row_block_inds)):
+            for j in range(len(self._col_block_inds)):
+                total += self._blocks[i, j].sq_resid_sum
+        return total
 
     @property
     def nbytes(self):
@@ -561,7 +1192,9 @@ class FormFactor2dTreeBlock(FormFactorBlockMatrix):
                 else:
                     spmat = spmat_par[row_inds, :][:, col_inds]
                 block = self.make_block(shape_model, I, J, spmat,
-                                        max_depth, force_max_depth)
+                                        max_depth, force_max_depth,
+                                        compression_type=self._root._compression_type,
+                                        compression_params=self._root._compression_params)
                 assert block is not None
                 row.append(block)
             blocks.append(row)
@@ -640,6 +1273,50 @@ class FormFactorQuadtreeBlock(FormFactor2dTreeBlock):
         PJ = P[:, :2] if J is None else P[J, :2]
         self._col_block_inds = [J for J in get_quadrant_order(PJ)]
 
+class FormFactorMinDepthQuadtreeBlock(FormFactorQuadtreeBlock):
+    """A form factor matrix block corresponding to a quadtree
+    partition which goes down at least min_depth levels. Only
+    for use as a root block.
+
+    """
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+
+    def _set_block_inds(self, shape_model, I, J):
+        P = shape_model.P
+
+        PI = P[:, :2]
+        prev_level_block_inds = [I for I in get_quadrant_order(PI)]
+
+        for _ in range(1, self._root._min_depth):
+            current_level_block_inds = []
+            for i in range(len(prev_level_block_inds)):
+                PI = P[prev_level_block_inds[i], :2]
+                current_level_block_inds += [prev_level_block_inds[i][I] for I in get_quadrant_order(PI)]
+            prev_level_block_inds = current_level_block_inds
+
+        self._row_block_inds = prev_level_block_inds
+        self._col_block_inds = prev_level_block_inds
+
+class FormFactorObbQuadtreeBlock(FormFactor2dTreeBlock):
+    """A form factor matrix block corresponding to one level of a quadtree
+    via oriented bounding box partitioning.
+
+    """
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+
+    def make_child_block(self, *args):
+        return self.root.make_obb_quadtree_block(*args)
+
+    def _set_block_inds(self, shape_model, I, J):
+
+        self._row_block_inds = get_obb_partition_2d(np.arange(shape_model.F.shape[0]) if I is None else I, shape_model)
+
+        self._col_block_inds = get_obb_partition_2d(np.arange(shape_model.F.shape[0]) if J is None else J, shape_model)
+
 class FormFactorOctreeBlock(FormFactor2dTreeBlock):
 
     def __init__(self, *args, **kwargs):
@@ -684,7 +1361,9 @@ class FormFactorPartitionBlock(FormFactorBlockMatrix):
             for j, J in enumerate(parts):
                 spmat = get_form_factor_matrix(shape_model, I, J)
                 block = self.make_block(shape_model, I, J, spmat,
-                                        max_depth, force_max_depth)
+                                        max_depth, force_max_depth,
+                                        compression_type=self._root._compression_type,
+                                        compression_params=self._root._compression_params)
                 assert block is not None
                 row_blocks.append(block)
             blocks.append(row_blocks)
@@ -708,8 +1387,8 @@ class CompressedFormFactorMatrix(scipy.sparse.linalg.LinearOperator):
     """
 
     def __init__(self, shape_model, tol=1e-5,
-                 min_size=16384, max_depth=None, force_max_depth=False,
-                 RootBlock=FormFactorQuadtreeBlock, **kwargs):
+                 min_size=16384, min_depth=None, max_depth=None, force_max_depth=False, RootBlock=FormFactorQuadtreeBlock,
+                 compression_type="svd", compression_params={}, add_residuals=False, truncated_sparse=False, **kwargs):
         """Create a new CompressedFormFactorMatrix.
 
         Parameters
@@ -723,6 +1402,9 @@ class CompressedFormFactorMatrix(scipy.sparse.linalg.LinearOperator):
         min_size : positive integer
             The minimum number of elements needed in a sub-block required
             before compression is attempted.
+        min_depth : None of positive integer
+            The minimum depth of the hieararchical matrix. If None is passed,
+            then the compression begins at the top level of the tree.
         max_depth : None or positive integer
             The maximum depth of the hierarchical matrix. If None is passed,
             then there is no maximum depth.
@@ -732,6 +1414,18 @@ class CompressedFormFactorMatrix(scipy.sparse.linalg.LinearOperator):
             same height. This is mostly useful for debugging. Default: False.
         RootBlock : class
             The class to use for the root node the hierarchical matrix.
+        compression_type : string
+            Indicator for the type of low-rank matrix decomposition to be used
+            in the algebraic compression.
+        compression_params: dictionary
+            Dictionary elements correspond to the parmaeters of the low-rank
+            decomposition, such as initial rank and number of power iterations.
+        add_residuals: boolean
+            Whether to add sparse residuals to the low-rank compression in
+            matrix blocks.
+        truncated_sparse: boolean
+            Whether to truncate the elements contained in the sparse matrix
+            blocks to meet the target error tolerance.
 
         """
         if tol < 0:
@@ -740,8 +1434,14 @@ class CompressedFormFactorMatrix(scipy.sparse.linalg.LinearOperator):
         self.shape_model = shape_model
         self._tol = tol
         self._min_size = min_size
+        self._min_depth = min_depth
         self._max_depth = max_depth
         self._force_max_depth = force_max_depth
+        self._compression_type = compression_type
+        self._compression_params = compression_params
+        self._add_residuals = add_residuals
+        self._truncated_sparse = truncated_sparse
+
         self._root = RootBlock(self, shape_model, max_depth=max_depth,
                                force_max_depth=force_max_depth, **kwargs)
 
@@ -785,15 +1485,36 @@ class CompressedFormFactorMatrix(scipy.sparse.linalg.LinearOperator):
 
     def make_sparse_block(self, *args, fmt='csr'):
         if fmt == 'csr':
-            return FormFactorCsrBlock(self, *args)
+            if self._truncated_sparse:
+                return FormFactorTruncatedCsrBlock(self, *args)
+            else:
+                return FormFactorCsrBlock(self, *args)
         else:
             raise Exception('unknown sparse matrix format "%s"' % fmt)
 
     def make_svd_block(self, *args):
         return FormFactorSvdBlock(self, *args)
 
+    def make_sparse_svd_block(self, *args):
+        return FormFactorSparseSvdBlock(self, *args)
+
+    def make_aca_block(self, *args):
+        return FormFactorAcaBlock(self, *args)
+
+    def make_sparse_aca_block(self, *args):
+        return FormFactorSparseAcaBlock(self, *args)
+
+    def make_id_block(self, *args):
+        return FormFactorIdBlock(self, *args)
+
+    def make_sparse_id_block(self, *args):
+        return FormFactorSparseIdBlock(self, *args)
+
     def make_quadtree_block(self, *args):
         return FormFactorQuadtreeBlock(self, *args)
+
+    def make_obb_quadtree_block(self, *args):
+        return FormFactorObbQuadtreeBlock(self, *args)
 
     def make_octree_block(self, *args):
         return FormFactorOctreeBlock(self, *args)
@@ -804,6 +1525,15 @@ class CompressedFormFactorMatrix(scipy.sparse.linalg.LinearOperator):
 
     def _matmat(self, x):
         return self._root@x
+
+    def __add__(self, x):
+        return self._root+x
+
+    def resid_frobenius_norm(self):
+        return np.power(self._root.sq_resid_sum, 0.5)
+
+    def full_mat_frobenius_norm(self):
+        return np.power(self._root.partial_norm, 0.5)
 
     def get_blocks_at_depth(self, depth):
         if depth > self.depth:
