+++ src/flux/linalg.py
@@ -1,6 +1,5 @@
 import numpy as np
 import scipy.sparse.linalg
-from sklearn.decomposition import NMF
 
 from flux.debug import DebugLinearOperator, IndentedPrinter
 from flux.util import nbytes
@@ -118,133 +117,6 @@ def estimate_sparsity_svd(spmat, tol, max_nbytes=None, k0=40):
 
         prev_k = k
         k *= 2
-
-
-
-
-# NMF ALGORITHMS
-
-def fit_nmf(spmat, k, max_iters=int(1e3), tol=1e-4, beta_loss=2, init='svd'):
-    if beta_loss == 2:
-        nmf = NMF(k,init='nndsvd' if init=='svd' else 'random', solver='cd', tol=tol, max_iter=max_iters, beta_loss=beta_loss)
-    elif beta_loss == 1:
-        nmf = NMF(k,init='nndsvdar' if init=='svd' else 'random', solver='mu', tol=tol, max_iter=max_iters, beta_loss=beta_loss)
-    W = nmf.fit_transform(spmat)
-    H = nmf.components_
-
-    if np.isnan(W).any() or np.isnan(H).any():
-        return None, None
-    
-    W = scipy.sparse.csr_matrix(W)
-    H = scipy.sparse.csr_matrix(H)
-    return W, H
-
-
-def estimate_rank_nmf(spmat, tol, max_nbytes=None, k0=40, max_iters=int(1e3), nmf_tol=1e-2, beta_loss=2):
-    assert tol < 1
-
-    if spmat.shape[0] == 0 or spmat.shape[1] == 0:
-        return 0
-
-    if spmat.shape == (1, 1):
-        return 1
-
-    m, k = min(spmat.shape), k0
-    num_iters = 0
-    while True:
-        num_iters += 1
-        k = min(k, m)
-        if not k >= 1:
-            raise RuntimeError('bad value of k')
-        W, H = fit_nmf(spmat, k, max_iters=max_iters, tol=nmf_tol, beta_loss=beta_loss)
-        nmf_nbytes = nbytes(W) + nbytes(H)
-        if max_nbytes is not None and nmf_nbytes >= max_nbytes:
-            return None
-        thresh = scipy.sparse.linalg.norm((W@H) - spmat, ord='fro') / scipy.sparse.linalg.norm(spmat, ord='fro')
-        if k == m or thresh <= tol:
-            return W, H, thresh
-        k *= 2
-
-
-def estimate_sparsity_nmf(spmat, tol, max_nbytes=None, k0=5, max_iters=int(1e3), nmf_tol=1e-2, beta_loss=2):
-    assert tol < 1
-
-    if spmat.shape[0] == 0 or spmat.shape[1] == 0:
-        return 0
-
-    if spmat.shape == (1, 1):
-        return 1
-
-    W, H = fit_nmf(spmat, k0, max_iters=max_iters, tol=nmf_tol, beta_loss=beta_loss, init='random')
-
-    if W is None:
-        return None
-    
-    resid = (spmat - (W@H)).A
-
-    num_resids = resid.shape[0] * resid.shape[1]
-    nnz_resid = np.count_nonzero(resid)
-
-    sorted_resid_idx = np.unravel_index(np.argsort(abs(resid), axis=None), resid.shape)
-
-    target = np.power(scipy.sparse.linalg.norm((W@H) - spmat, ord='fro'), 2) - np.power(tol*scipy.sparse.linalg.norm(spmat, ord='fro'), 2)
-
-    if target <= 0:
-        Sr = np.zeros_like(resid)
-        Sr = scipy.sparse.csr_matrix(Sr)
-
-    else:
-        cumulative_residual = np.cumsum(np.power(resid[sorted_resid_idx[0], sorted_resid_idx[1]][::-1], 2))
-        keep_resids = (cumulative_residual > target).nonzero()[0][0] + 1
-
-        Sr = np.copy(resid)
-        Sr[sorted_resid_idx[0][:num_resids-keep_resids], sorted_resid_idx[1][:num_resids-keep_resids]] = 0.
-        Sr = scipy.sparse.csr_matrix(Sr)
-
-    sparse_nmf_nbytes = nbytes(W) + nbytes(H) + nbytes(Sr)
-    if max_nbytes is not None and sparse_nmf_nbytes >= max_nbytes:
-        return None
-
-    return W, H, Sr
-
-
-def estimate_sparsity_nmf_weighted(spmat, FF_weights, tol, max_nbytes=None, k0=5, max_iters=int(1e3), nmf_tol=1e-2, beta_loss=2):
-    assert tol < 1
-
-    if spmat.shape[0] == 0 or spmat.shape[1] == 0:
-        return 0
-
-    if spmat.shape == (1, 1):
-        return 1
-
-    W, H = fit_nmf(spmat, k, max_iters=max_iters, tol=nmf_tol, beta_loss=beta_loss)
-    resid = (spmat - (W@H)).A
-
-    num_resids = resid.shape[0] * resid.shape[1]
-    nnz_resid = np.count_nonzero(resid)
-
-    weighted_resid = np.multiply(FF_weights, resid)
-    sorted_weighted_resid_idx = np.unravel_index(np.argsort(abs(weighted_resid), axis=None), resid.shape)
-
-    m, s = min(spmat.shape), s0
-    num_iters = 0
-    while True:
-        num_iters += 1
-        s = min(s, nnz_resid)
-        if not s >= 1:
-            raise RuntimeError('bad value of s')
-
-        Sr = np.copy(resid)
-        Sr[sorted_weighted_resid_idx[0][:num_resids-s], sorted_weighted_resid_idx[1][:num_resids-s]] = 0.
-        Sr = scipy.sparse.csr_matrix(Sr)
-        
-        sparse_nmf_nbytes = nbytes(W) + nbytes(H) + nbytes(Sr)
-        if max_nbytes is not None and sparse_nmf_nbytes >= max_nbytes:
-            return None
-        thresh = scipy.sparse.linalg.norm(((W@H) + Sr) - spmat, ord='fro') / scipy.sparse.linalg.norm(spmat, ord='fro')
-        if s == nnz_resid or thresh <= tol:
-            return W, H, Sr, thresh
-        s *= 2
 
 
 
@@ -517,115 +389,6 @@ def estimate_sparsity_partial_aca(spmat, tol, max_nbytes=None, k0=40):
 
 
 
-# BRP ALGORITHMS
-
-def bilateral_random_projection(X, k):
-    
-    m, n = X.shape
-    
-    A1 = np.random.normal(loc=0.0, scale=1.0, size=(n,k))
-    Y1 = X @ A1
-    
-    A2 = np.copy(Y1)
-    Y2 = X.T @ A2
-    
-    A1 = np.copy(Y2)
-    Y1 = X @ A1
-
-    return scipy.sparse.csr_matrix(Y1), scipy.sparse.csr_matrix(np.linalg.inv(A2.T @ Y1)), scipy.sparse.csr_matrix(Y2)
-
-
-def estimate_rank_brp(spmat, tol, max_nbytes=None, k0=40):
-    assert tol < 1
-
-    if spmat.shape[0] == 0 or spmat.shape[1] == 0:
-        return 0
-
-    if spmat.shape == (1, 1):
-        return 1
-
-    m, k = min(spmat.shape), k0
-    num_iters = 0
-    while True:
-        num_iters += 1
-        k = min(k, m)
-        if not k >= 1:
-            raise RuntimeError('bad value of k')
-        Y1, D, Y2 = bilateral_random_projection(spmat.A, k)
-        brp_nbytes = nbytes(Y1) + nbytes(D) + nbytes(Y2)
-        if max_nbytes is not None and brp_nbytes >= max_nbytes:
-            return None
-        thresh = scipy.sparse.linalg.norm((Y1 @ D @ Y2.T) - spmat, ord='fro') / scipy.sparse.linalg.norm(spmat, ord='fro')
-        if k == m or thresh <= tol:
-            return Y1, D, Y2, thresh
-        k += 5
-
-
-def estimate_sparsity_brp(spmat, tol, max_nbytes=None, k0=5):
-    assert tol < 1
-
-    if spmat.shape[0] == 0 or spmat.shape[1] == 0:
-        return 0
-
-    if spmat.shape == (1, 1):
-        return 1
-
-    prev_k = 0
-    prev_nbytes = np.inf
-    Sr_prev = np.zeros(spmat.shape)
-    
-    m, k = min(spmat.shape), k0
-    while True:
-        k = min(k, m)
-        
-        if not k >= 1:
-            raise RuntimeError('bad value of k')
-
-        Y1, D, Y2 = bilateral_random_projection(spmat.A, k)
-        
-        resid = (spmat - (Y1 @ D @ Y2.T)).A
-        
-        if k == m:
-            return Y1, D, Y2, scipy.sparse.csr_matrix(np.zeros_like(resid))
-
-        num_resids = resid.shape[0] * resid.shape[1]
-        nnz_resid = np.count_nonzero(resid)
-
-        sorted_resid_idx = np.unravel_index(np.argsort(abs(resid), axis=None), resid.shape)
-
-        target = np.power(np.linalg.norm(resid, ord='fro'), 2) - np.power(tol*scipy.sparse.linalg.norm(spmat, ord='fro'), 2)
-
-        if target <= 0:
-            Sr = np.zeros_like(resid)
-            Sr = scipy.sparse.csr_matrix(Sr)
-
-        else:
-            cumulative_residual = np.cumsum(np.power(resid[sorted_resid_idx[0], sorted_resid_idx[1]][::-1], 2))
-            keep_resids = (cumulative_residual > target).nonzero()[0][0] + 1
-
-            Sr = np.copy(resid)
-            Sr[sorted_resid_idx[0][:num_resids-keep_resids], sorted_resid_idx[1][:num_resids-keep_resids]] = 0.
-            Sr = scipy.sparse.csr_matrix(Sr)
-
-        sparse_brp_nbytes = nbytes(Y1) + nbytes(D) + nbytes(Y2) + nbytes(Sr)
-
-        if sparse_brp_nbytes >= prev_nbytes:
-            return Y1_prev, D_prev, Y2_prev, Sr_prev
-
-        if max_nbytes is not None and sparse_brp_nbytes >= max_nbytes:
-            return None
-
-        prev_nbytes = sparse_brp_nbytes
-        Y1_prev = Y1
-        D_prev = D
-        Y2_prev = Y2
-        Sr_prev = Sr
-        prev_k = k
-        k += 5
-
-
-
-
 # RANDOM ID ALGORITHMS
 
 '''
@@ -752,56 +515,6 @@ def estimate_sparsity_random_id(spmat, tol, max_nbytes=None, k0=40, p=5, q=5):
 
 # OTHER RANDOM ALGORITHMS
 
-def randomized_HALS(X, Q, k, tol=1e-4, max_iters=1e4):
-    m,n = X.shape
-    
-    l = Q.shape[1]
-    
-    W = np.random.uniform(low=0.0, high=1.0, size=(m,k))
-    _W = np.random.uniform(low=0.0, high=1.0, size=(l,k))
-    H = np.random.uniform(low=0.0, high=1.0, size=(k,n))
-    
-    B = Q.T @ X        # l x n
-    
-    criterion = True
-    num_iters = 0
-    H_norm = np.linalg.norm(H, ord='fro')
-    while criterion:
-        
-        R = B.T @ _W       # n x k
-        S = _W.T @ _W      # k x k
-        for j in range(k):
-            if S[j,j] != 0:
-                H[j,:] = H[j,:] + ((R[:,j] - (H.T @ S[:,j])) / S[j,j])
-                H[j,:] = np.maximum(H[j,:], 0.)
-            
-        T = B @ H.T        # l x k
-        V = H @ H.T        # k x k
-        for j in range(k):
-            if V[j,j] != 0:
-                _W[:,j] = _W[:,j] + ((T[:,j] - (_W @ V[:,j])) / V[j,j])
-                W[:,j] = np.maximum(0, Q @ _W[:,j])
-                _W[:,j] = Q.T @ W[:,j]
-            
-        num_iters += 1
-        if num_iters >= max_iters:
-            print('Exceeded max_iters')
-            criterion = False
-           
-        new_H_norm = np.linalg.norm(H, ord='fro')
-        if (abs(H_norm - new_H_norm) / H_norm)  < tol:
-            criterion = False
-        
-        H_norm = new_H_norm
-
-        if np.isnan(W).any() or np.isnan(H).any():
-            return None, None
-            
-    W = scipy.sparse.csr_matrix(W)
-    H = scipy.sparse.csr_matrix(H)
-    return W, H
-
-
 def randomized_svd(B, Q, k):
     _U, S, Vt = np.linalg.svd(B, full_matrices=False)
     U = Q @ _U
@@ -831,49 +544,6 @@ def simple_range_finder(X, k, p=5, q=1, rand_dist='uniform'):
     Q = np.linalg.qr(Y, mode='reduced')[0]
     
     return Q
-
-
-def estimate_sparsity_random_nmf(spmat, tol, max_nbytes=None, k0=5, max_iters=int(1e3), nmf_tol=1e-2, p=5, q=1):
-    assert tol < 1
-
-    if spmat.shape[0] == 0 or spmat.shape[1] == 0:
-        return 0
-
-    if spmat.shape == (1, 1):
-        return 1
-
-    Q = simple_range_finder(spmat, k0, p=p, q=q)
-    W, H = randomized_HALS(spmat, Q, k0, tol=nmf_tol, max_iters=max_iters)
-
-    if W is None:
-        return None
-    
-    resid = (spmat - (W@H)).A
-
-    num_resids = resid.shape[0] * resid.shape[1]
-    nnz_resid = np.count_nonzero(resid)
-
-    sorted_resid_idx = np.unravel_index(np.argsort(abs(resid), axis=None), resid.shape)
-
-    target = np.power(scipy.sparse.linalg.norm((W@H) - spmat, ord='fro'), 2) - np.power(tol*scipy.sparse.linalg.norm(spmat, ord='fro'), 2)
-
-    if target <= 0:
-        Sr = np.zeros_like(resid)
-        Sr = scipy.sparse.csr_matrix(Sr)
-
-    else:
-        cumulative_residual = np.cumsum(np.power(resid[sorted_resid_idx[0], sorted_resid_idx[1]][::-1], 2))
-        keep_resids = (cumulative_residual > target).nonzero()[0][0] + 1
-
-        Sr = np.copy(resid)
-        Sr[sorted_resid_idx[0][:num_resids-keep_resids], sorted_resid_idx[1][:num_resids-keep_resids]] = 0.
-        Sr = scipy.sparse.csr_matrix(Sr)
-
-    sparse_nmf_nbytes = nbytes(W) + nbytes(H) + nbytes(Sr)
-    if max_nbytes is not None and sparse_nmf_nbytes >= max_nbytes:
-        return None
-
-    return W, H, Sr
 
 
 def estimate_rank_random_svd(spmat, tol, max_nbytes=None, k0=40, p=5, q=1):
